feat: Create the first rule 'no-at-important'
This rule disallow the use of `!important`. The rule accept an option `ignoreLayers` which permit to accept the use of `!important` in some layers.
feat: Create the first rule 'no-at-important'

This rule disallow the use of `!important`. The rule accept an option `ignoreLayers` which permit to accept the use of `!important` in some layers.